Migrate Mutation.updateMe
